Labs/requirements.txt
psycopg2==2.9.5
cassandra-driver==3.25.0
simplejson==3.17.2